requirements.txt
altair==4.0
Pillow>=8.2.0
piexif>=1.1.3
watchdog>=2.1.2
streamlit==1.24.0
starlette>=0.18.0
python-magic>=0.4.24
requests>=2.26.0
python-multipart
jinja2>=3.0.0
pandas
plotly
google-play-scraper
openai
PyYAML
pycountry
app-store-scraper
itunes-app-scraper-dmi